fix: auth guard
Se corrigió un error en el auth guard que no verificaba el token

+++ src/modules/vehicles/services/vehicle.service.ts
@@ -56,7 +56,11 @@ export class VehicleService {
         return 'Vehicle saved succesfully';
     }
 
-    public async editVehicle(id: ObjectId, vehicle: EditVehicleDto): Promise<string> {
+    public async editVehicle(id: ObjectId, vehicle: EditVehicleDto, uid: string): Promise<string> {
+        const car: Vehicle = await this.findOneVehicle(id);
+
+        if (car.owner !== uid) throw new HttpException("You don't have permissions to do this action", HttpStatus.UNAUTHORIZED);
+
         const res: UpdateResult = await this.vehicleRepositoy.update(id, vehicle);
 
         if (res.affected === 0) throw new HttpException('Vehicle not found', HttpStatus.NOT_FOUND);
@@ -64,13 +68,17 @@ export class VehicleService {
         return 'Changes saved succesfully';
     }
 
-    public async deleteVehcile(id: ObjectId): Promise<void> {
+    public async deleteVehcile(id: ObjectId, uid: string): Promise<void> {
         const vehicle: Vehicle = await this.findOneVehicle(id);
+
+        if (vehicle.owner !== uid) throw new HttpException("You don't have permissions to do this action", HttpStatus.UNAUTHORIZED);
+
         const res: DeleteResult = await this.vehicleRepositoy.delete(new ObjectId(id));
 
         if (res.affected === 0) throw new HttpException('Oops!, something went wrong', HttpStatus.INTERNAL_SERVER_ERROR);
         if (res.affected === 1) {
             vehicle.images.map(image => this.cloudinaryService.deleteFile(image.id));
+
             throw new HttpException('User deleted', HttpStatus.OK);
         };
     }
